+++ src/app/main/apps/processo/processo-view/vinculacao-documento/vinculacao-documento.component.ts
@@ -32,6 +32,8 @@ export class VinculacaoDocumentoComponent implements OnInit, OnDestroy {
 
     juntada$: Observable<Juntada>;
     juntada: Juntada;
+    juntadaVinculada$: Observable<Juntada>;
+    juntadaVinculada: Juntada;
 
     isSaving$: Observable<boolean>;
     errors$: Observable<any>;
@@ -57,6 +59,7 @@ export class VinculacaoDocumentoComponent implements OnInit, OnDestroy {
         this.isSaving$ = this._store.pipe(select(fromStore.getIsSaving));
         this.errors$ = this._store.pipe(select(fromStore.getErrors));
         this.juntada$ = this._store.pipe(select(fromStore.getJuntada));
+        this.juntadaVinculada$ = this._store.pipe(select(fromStore.getJuntadaVinculada));
 
         this.modalidadeVinculacaoDocumentoPagination = new Pagination();
         this.documentoVinculadoPagination = new Pagination();
@@ -103,6 +106,15 @@ export class VinculacaoDocumentoComponent implements OnInit, OnDestroy {
                 this._changeDetectorRef.detectChanges();
             }
         );
+
+        this.juntadaVinculada$.pipe(
+            filter(juntada => !!juntada)
+        ).subscribe(
+            (juntada) => {
+                this.juntadaVinculada = juntada;
+                this._changeDetectorRef.detectChanges();
+            }
+        );
     }
 
     doAbort(): void {
@@ -130,9 +142,11 @@ export class VinculacaoDocumentoComponent implements OnInit, OnDestroy {
         );
 
         vinculacaoDocumento.documento = this.juntada.documento;
+        if (this.juntadaVinculada) {
+            vinculacaoDocumento.documentoVinculado = this.juntadaVinculada.documento;
+        }
 
         this._store.dispatch(new fromStore.SaveVinculacaoDocumento(vinculacaoDocumento));
 
     }
-
 }